Extract some common code into abstract macros
These can then be implemented for custom data sources.

Fixes: SE-13745
