+++ src/main/java/sirius/pasta/noodle/macros/InlineSvgResourceMacro.java
@@ -12,70 +12,39 @@ import org.w3c.dom.Document;
 import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
 import sirius.kernel.tokenizer.Position;
-import sirius.pasta.noodle.Environment;
 import sirius.pasta.noodle.compiler.CompilationContext;
-import sirius.pasta.noodle.compiler.ir.Node;
 import sirius.web.resources.Resource;
 import sirius.web.resources.Resources;
 
 import javax.annotation.Nonnull;
-import java.util.List;
 
 /**
- * Provides a macro for inlining an SVG file into a DOM tree by dropping the XML declaration from the beginning and
+ * Provides a macro for inlining an SVG resource into a DOM tree by dropping the XML declaration from the beginning and
  * returning the rest.
  */
 @Register
-public class InlineSvgResourceMacro extends XmlProcessingMacro {
+public class InlineSvgResourceMacro extends InlineSvgMacro {
 
     @Part
     private Resources resources;
 
     @Override
-    protected Class<?> getType() {
-        return String.class;
-    }
-
-    @Override
-    protected void verifyArguments(CompilationContext compilationContext, Position position, List<Class<?>> args) {
-        if (args.size() != 1 || !CompilationContext.isAssignableTo(args.getFirst(), String.class)) {
-            throw new IllegalArgumentException("Expected a single String as argument.");
+    protected void verifyPath(CompilationContext context, Position position, String path) {
+        if (resources.resolve(path).isEmpty()) {
+            context.warning(position, "Unknown resource: %s", path);
         }
     }
 
     @Override
-    public void verify(CompilationContext context, Position position, List<Node> args) {
-        super.verify(context, position, args);
-
-        if (args.getFirst().isConstant()) {
-            String resourceName = String.valueOf(args.getFirst().getConstantValue());
-            if (resources.resolve(resourceName).isEmpty()) {
-                context.warning(position, "Unknown resource: %s", resourceName);
-            }
-        }
-    }
-
-    @Override
-    public boolean isConstant(CompilationContext context, List<Node> args) {
-        return args.getFirst().isConstant();
-    }
-
-    @Override
-    public Object invoke(Environment environment, Object[] args) {
-        String path = (String) args[0];
-        return processResource(path);
-    }
-
-    @Nonnull
-    private String processResource(String path) {
+    protected Document getSvgDocument(String path) {
         if (!path.startsWith("/assets")) {
             throw new IllegalArgumentException("Only assets can be inlined for security reasons.");
         }
 
         Resource resource =
                 resources.resolve(path).orElseThrow(() -> new IllegalArgumentException("Unknown resource: " + path));
-        Document document = parseDocument(resource, "svg");
-        return stringifyElement(document.getDocumentElement(), false);
+
+        return parseDocument(resource, "svg");
     }
 
     @Nonnull
@@ -87,6 +56,6 @@ public class InlineSvgResourceMacro extends XmlProcessingMacro {
     @Nonnull
     @Override
     public String getDescription() {
-        return "Returns the root <svg> tag of an SVG file as string, to be included in other DOM trees.";
+        return "Returns the root <svg> tag of an SVG resource as string, to be included in other DOM trees.";
     }
 }
